# openapi.yaml
openapi: 3.1.0
info:
  title: Qarax API
  description: REST API for managing virtual machines and hypervisor hosts
  license:
    name: ''
  version: 0.1.0
paths:
...
  /vms:
    get:
      tags:
      - vms
      operationId: list
      parameters:
      - name: name
        in: query
        description: Optional name filter
        required: false
        schema:
          type:
          - string
          - 'null'
      - name: tags
        in: query
        description: Comma-separated list of tags; returned VMs must have all specified tags
        required: false
        schema:
          type:
          - string
          - 'null'
      responses:
        '200':
          description: List all VMs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Vm'
        '500':
          description: Internal server error
    post:
      tags:
      - vms
      operationId: create
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewVm'
        required: true
      responses:
        '201':
          description: VM created successfully (synchronous)
          content:
            application/json:
              schema:
                type: string
        '202':
          description: VM creation started asynchronously
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateVmResponse'
        '422':
          description: Invalid input
        '500':
          description: Internal server error
  /vms/preflight:
    post:
      tags:
      - vms
      operationId: preflight_image
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VmImagePreflightRequest'
        required: true
      responses:
        '200':
          description: OCI image preflight completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VmImagePreflightResponse'
        '422':
          description: Image or host is not preflightable
        '500':
          description: Internal server error
  /vms/{vm_id}:
    get:
      tags:
      - vms
      operationId: get
      parameters:
      - name: vm_id
        in: path
        description: VM unique identifier
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: VM details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vm'
        '404':
          description: VM not found
        '500':
          description: Internal server error
    delete:
      tags:
      - vms
      operationId: delete
      parameters:
      - name: vm_id
        in: path
        description: VM unique identifier
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '204':
          description: VM deleted successfully
        '404':
          description: VM not found
        '500':
          description: Internal server error
  /vms/{vm_id}/commit:
    post:
      tags:
      - vms
      operationId: commit
      parameters:
      - name: vm_id
        in: path
        description: VM unique identifier
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommitVmRequest'
        required: true
      responses:
        '202':
          description: Commit job accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommitVmResponse'
        '404':
          description: VM not found
        '422':
          description: VM not eligible for commit
  /vms/{vm_id}/console:
    get:
      tags:
      - vms
      operationId: console_log
      parameters:
      - name: vm_id
        in: path
        description: VM unique identifier
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Console log content
          content:
            text/plain:
              schema:
                type: string
        '404':
          description: Console logging not available for this VM
        '500':
          description: Internal server error
...
  /vms/{vm_id}/disks:
    post:
      tags:
      - vms
      summary: |-
        Attach a storage object to a VM as a disk.
        For VMs in `Created` state the disk is only recorded in the database and will be
        passed to Cloud Hypervisor when the VM starts.  For VMs in `Running` or `Shutdown`
        state the disk is recorded **and** immediately applied via the Cloud Hypervisor API
        (CH keeps the VM definition after shutdown, so disk changes are accepted).
      operationId: attach_disk
      parameters:
      - name: vm_id
        in: path
        description: VM unique identifier
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AttachDiskRequest'
        required: true
      responses:
        '201':
          description: Disk attached (and applied to CH if VM is running or shutdown)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VmDisk'
        '404':
          description: VM or storage object not found
        '422':
          description: VM not in Created, Running, or Shutdown state
        '500':
          description: Internal server error
  /vms/{vm_id}/disks/{device_id}:
    delete:
      tags:
      - vms
      operationId: remove_disk
      parameters:
      - name: vm_id
        in: path
        description: VM unique identifier
        required: true
        schema:
          type: string
          format: uuid
      - name: device_id
        in: path
        description: Disk logical name (e.g. "disk0")
        required: true
        schema:
          type: string
      responses:
        '204':
          description: Disk removed (and removed from CH if VM was running or shutdown)
        '404':
          description: VM or disk not found
        '422':
          description: VM not in Created, Running, or Shutdown state
        '500':
          description: Internal server error
  /vms/{vm_id}/disks/{disk_id}/resize:
    put:
      tags:
      - vms
      operationId: resize_disk
      parameters:
      - name: vm_id
        in: path
        description: VM unique identifier
        required: true
        schema:
          type: string
          format: uuid
      - name: disk_id
        in: path
        description: Logical disk name (e.g. "rootfs", "disk0")
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DiskResizeRequest'
        required: true
      responses:
        '200':
          description: Disk resized successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StorageObject'
        '404':
          description: VM or disk not found
        '422':
          description: VM not stopped, disk not resizable, or size invalid
        '500':
          description: Internal server error
  /vms/{vm_id}/force-stop:
    post:
      tags:
      - vms
      operationId: force_stop
      parameters:
      - name: vm_id
        in: path
        description: VM unique identifier
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: VM force stopped successfully
        '404':
          description: VM not found
        '500':
          description: Internal server error
...
  /vms/{vm_id}/migrate:
    post:
      tags:
      - vms
      operationId: migrate
      parameters:
      - name: vm_id
        in: path
        description: VM unique identifier
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VmMigrateRequest'
        required: true
      responses:
        '202':
          description: Migration accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VmMigrateResponse'
        '404':
          description: VM or host not found
        '422':
          description: VM not in a migratable state
        '500':
          description: Internal server error
  /vms/{vm_id}/nics:
    get:
      tags:
      - vms
      operationId: list_nics
      parameters:
      - name: vm_id
        in: path
        description: VM unique identifier
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: List of NICs attached to the VM
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/NetworkInterface'
        '404':
          description: VM not found
        '500':
          description: Internal server error
    post:
      tags:
      - vms
      operationId: add_nic
      parameters:
      - name: vm_id
        in: path
        description: VM unique identifier
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewVmNetwork'
        required: true
      responses:
        '201':
          description: NIC added (and hotplugged if VM is running)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NetworkInterface'
        '404':
          description: VM not found
        '409':
          description: Device ID already in use
        '422':
          description: VM not in Created or Running state
        '500':
          description: Internal server error
  /vms/{vm_id}/nics/{device_id}:
    delete:
      tags:
      - vms
      operationId: remove_nic
      parameters:
      - name: vm_id
        in: path
        description: VM unique identifier
        required: true
        schema:
          type: string
          format: uuid
      - name: device_id
        in: path
        description: NIC device ID (e.g. "net0")
        required: true
        schema:
          type: string
      responses:
        '204':
          description: NIC removed (and hotunplugged if VM was running)
        '404':
          description: VM or NIC not found
        '422':
          description: VM not in Created or Running state
        '500':
          description: Internal server error
  /vms/{vm_id}/pause:
    post:
      tags:
      - vms
      operationId: pause
      parameters:
      - name: vm_id
        in: path
        description: VM unique identifier
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: VM paused successfully
        '404':
          description: VM not found
        '500':
          description: Internal server error
  /vms/{vm_id}/resize:
    put:
      tags:
      - vms
      operationId: resize_vm
      parameters:
      - name: vm_id
        in: path
        description: VM unique identifier
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VmResizeRequest'
        required: true
      responses:
        '200':
          description: VM resized successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vm'
        '404':
          description: VM not found
        '422':
          description: VM not running, or resize parameters out of range
        '500':
          description: Internal server error
  /vms/{vm_id}/restore:
    post:
      tags:
      - vms
      operationId: restore
      parameters:
      - name: vm_id
        in: path
        description: VM unique identifier
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RestoreRequest'
        required: true
      responses:
        '200':
          description: VM restored from snapshot
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vm'
        '404':
          description: VM or snapshot not found
        '422':
          description: VM not in a restoreable state or snapshot not ready
        '500':
          description: Internal server error
  /vms/{vm_id}/resume:
    post:
      tags:
      - vms
      operationId: resume
      parameters:
      - name: vm_id
        in: path
        description: VM unique identifier
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: VM resumed successfully
        '404':
          description: VM not found
        '500':
          description: Internal server error
  /vms/{vm_id}/security-groups:
    get:
      tags:
      - vms
      operationId: list_security_groups
      parameters:
      - name: vm_id
        in: path
        description: VM unique identifier
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: List security groups bound to the VM
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SecurityGroup'
        '404':
          description: VM not found
        '500':
          description: Internal server error
    post:
      tags:
      - vms
      operationId: attach_security_group
      parameters:
      - name: vm_id
        in: path
        description: VM unique identifier
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AttachSecurityGroupRequest'
        required: true
      responses:
        '204':
          description: Security group bound to VM
        '404':
          description: VM or security group not found
        '500':
          description: Internal server error
...
  /vms/{vm_id}/snapshots:
    get:
      tags:
      - vms
      operationId: list_snapshots
      parameters:
      - name: vm_id
        in: path
        description: VM unique identifier
        required: true
        schema:
          type: string
          format: uuid
      - name: name
        in: query
        description: Optional name filter for list queries
        required: false
        schema:
          type:
          - string
          - 'null'
      responses:
        '200':
          description: List snapshots for a VM
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Snapshot'
        '404':
          description: VM not found
        '500':
          description: Internal server error
    post:
      tags:
      - vms
      operationId: create_snapshot
      parameters:
      - name: vm_id
        in: path
        description: VM unique identifier
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSnapshotRequest'
        required: true
      responses:
        '201':
          description: Snapshot created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Snapshot'
        '404':
          description: VM not found
        '500':
          description: Internal server error
  /vms/{vm_id}/start:
    post:
      tags:
      - vms
      operationId: start
      parameters:
      - name: vm_id
        in: path
        description: VM unique identifier
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '202':
          description: VM start accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VmStartResponse'
        '404':
          description: VM not found
        '422':
          description: VM not in a startable state
        '500':
          description: Internal server error
  /vms/{vm_id}/stop:
    post:
      tags:
      - vms
      operationId: stop
      parameters:
      - name: vm_id
        in: path
        description: VM unique identifier
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: VM stopped successfully
        '404':
          description: VM not found
        '500':
          description: Internal server error
...
components:
  schemas:
...
    AttachDiskRequest:
      type: object
      required:
      - storage_object_id
      properties:
        boot_order:
          type:
          - integer
          - 'null'
          format: int32
          description: 'Boot priority — lower is higher priority (default: `0`).'
        logical_name:
          type:
          - string
          - 'null'
          description: Logical device name inside the VM (e.g. "vda"); auto-generated if omitted.
        storage_object_id:
          type: string
          format: uuid
          description: 'Storage object ID (any type: disk, oci_image, iso, etc.).'
...
    CommitVmRequest:
      type: object
      required:
      - storage_pool_id
      - size_bytes
      properties:
        size_bytes:
          type: integer
          format: int64
          description: Size of the committed disk in bytes.
        storage_pool_id:
          type: string
          format: uuid
          description: Storage pool to create the raw disk on (must be Local or NFS, attached to the VM's host).
    CommitVmResponse:
      type: object
      required:
      - vm_id
      - job_id
      properties:
        job_id:
          type: string
          format: uuid
        vm_id:
          type: string
          format: uuid
...
    CreateVmResponse:
      type: object
      required:
      - vm_id
      - job_id
      properties:
        job_id:
          type: string
          format: uuid
        vm_id:
          type: string
          format: uuid
...
    DiskResizeRequest:
      type: object
      description: Request body for `PUT /vms/{vm_id}/disks/{disk_id}/resize`.
      required:
      - new_size_bytes
      properties:
        new_size_bytes:
          type: integer
          format: int64
          description: New disk size in bytes. Must be larger than the current size and a multiple of 1 MiB.
...
    NewVm:
      type: object
      required:
      - name
      properties:
        accelerator_config:
          description: |-
            Accelerator (GPU) configuration. When set, GPU-aware scheduling picks a
            host with available GPUs matching these filters, and VFIO passthrough
            devices are attached to the VM.
        architecture:
          type:
          - string
          - 'null'
        boot_mode:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/BootMode'
        boot_source_id:
          type:
          - string
          - 'null'
          format: uuid
        boot_vcpus:
          type:
          - integer
          - 'null'
          format: int32
        cloud_init_meta_data:
          type:
          - string
          - 'null'
          description: Cloud-init meta-data (raw YAML). Auto-generated from vm id/name if omitted.
        cloud_init_network_config:
          type:
          - string
          - 'null'
          description: |-
            Cloud-init network-config (raw YAML). When provided, kernel `ip=` cmdline
            params are suppressed so cloud-init owns networking.
        cloud_init_user_data:
          type:
          - string
          - 'null'
          description: |-
            Cloud-init user-data (raw YAML). When provided a NoCloud seed image is
            generated and attached as a read-only disk to the VM.
        config: {}
        cpu_topology: {}
        description:
          type:
          - string
          - 'null'
        hypervisor:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/Hypervisor'
        image_ref:
          type:
          - string
          - 'null'
          description: |-
            OCI image reference to use as root filesystem (e.g. "docker.io/library/ubuntu:22.04").
            When set, the handler will check whether the selected host has an OverlayBD storage pool
            and the image is served via lazy block loading (virtio-blk).
        instance_type_id:
          type:
          - string
          - 'null'
          format: uuid
        kvm_hyperv:
          type:
          - boolean
          - 'null'
        max_vcpus:
          type:
          - integer
          - 'null'
          format: int32
        memory_hotplug_size:
          type:
          - integer
          - 'null'
          format: int64
        memory_hugepage_size:
          type:
          - integer
          - 'null'
          format: int64
        memory_hugepages:
          type:
          - boolean
          - 'null'
        memory_mergeable:
          type:
          - boolean
          - 'null'
        memory_prefault:
          type:
          - boolean
          - 'null'
        memory_shared:
          type:
          - boolean
          - 'null'
        memory_size:
          type:
          - integer
          - 'null'
          format: int64
        memory_thp:
          type:
          - boolean
          - 'null'
        name:
          type: string
        network_id:
          type:
          - string
          - 'null'
          format: uuid
          description: Network ID to attach the VM to (triggers IPAM allocation).
        networks:
          type:
          - array
          - 'null'
          items:
            $ref: '#/components/schemas/NewVmNetwork'
          description: Optional network interfaces to attach at create time (passed to qarax-node).
        numa_config:
          description: |-
            NUMA configuration. When set, the VM is pinned to the specified NUMA node.
            If accelerator_config has prefer_local_numa=true (the default), GPU-local NUMA
            is used instead and this field is ignored.
        persistent_upper_pool_id:
          type:
          - string
          - 'null'
          format: uuid
          description: |-
            When set alongside `image_ref`, the OverlayBD upper layer (upper.data +
            upper.index) is stored as a persistent `OverlaybdUpper` StorageObject on
            this pool instead of being ephemeral. The pool must be Local or NFS and
            must be attached to the host running the VM.
        root_disk_object_id:
          type:
          - string
          - 'null'
          format: uuid
        security_group_ids:
          type:
          - array
          - 'null'
          items:
            type: string
          description: |-
            Security groups to bind to the VM. Rules apply to managed routed traffic
            on every managed NIC attached to the VM.
        tags:
          type:
          - array
          - 'null'
          items:
            type: string
        vm_template_id:
          type:
          - string
          - 'null'
          format: uuid
...
    StorageObject:
      type: object
      required:
      - id
      - name
      - storage_pool_id
      - object_type
      - size_bytes
      - config
      properties:
        config: {}
        id:
          type: string
          format: uuid
        name:
          type: string
        object_type:
          $ref: '#/components/schemas/StorageObjectType'
        parent_id:
          type:
          - string
          - 'null'
          format: uuid
        size_bytes:
          type: integer
          format: int64
        storage_pool_id:
          type: string
          format: uuid
...
    Vm:
      type: object
      required:
      - id
      - name
      - tags
      - status
      - hypervisor
      - boot_mode
      - boot_vcpus
      - max_vcpus
      - kvm_hyperv
      - memory_size
      - memory_mergeable
      - memory_shared
      - memory_hugepages
      - memory_prefault
      - memory_thp
      - config
      properties:
        boot_mode:
          $ref: '#/components/schemas/BootMode'
        boot_source_id:
          type:
          - string
          - 'null'
          format: uuid
        boot_vcpus:
          type: integer
          format: int32
        cloud_init_meta_data:
          type:
          - string
          - 'null'
        cloud_init_network_config:
          type:
          - string
          - 'null'
        cloud_init_user_data:
          type:
          - string
          - 'null'
        config: {}
        cpu_topology: {}
        description:
          type:
          - string
          - 'null'
        host_id:
          type:
          - string
          - 'null'
          format: uuid
        hypervisor:
          $ref: '#/components/schemas/Hypervisor'
        id:
          type: string
          format: uuid
        image_ref:
          type:
          - string
          - 'null'
        kvm_hyperv:
          type: boolean
        max_vcpus:
          type: integer
          format: int32
        memory_hotplug_size:
          type:
          - integer
          - 'null'
          format: int64
        memory_hugepage_size:
          type:
          - integer
          - 'null'
          format: int64
        memory_hugepages:
          type: boolean
        memory_mergeable:
          type: boolean
        memory_prefault:
          type: boolean
        memory_shared:
          type: boolean
        memory_size:
          type: integer
          format: int64
        memory_thp:
          type: boolean
        name:
          type: string
        status:
          $ref: '#/components/schemas/VmStatus'
        tags:
          type: array
          items:
            type: string
    VmDisk:
      type: object
      required:
      - id
      - vm_id
      - logical_name
      - device_path
      - read_only
      - direct
      - vhost_user
      - num_queues
      - queue_size
      - pci_segment
      - config
      properties:
        boot_order:
          type:
          - integer
          - 'null'
          format: int32
        config: {}
        device_path:
          type: string
        direct:
          type: boolean
        id:
          type: string
          format: uuid
        logical_name:
          type: string
        num_queues:
          type: integer
          format: int32
        pci_segment:
          type: integer
          format: int32
        queue_size:
          type: integer
          format: int32
        rate_limit_group:
          type:
          - string
          - 'null'
        rate_limiter: {}
        read_only:
          type: boolean
        serial_number:
          type:
          - string
          - 'null'
        storage_object_id:
          type:
          - string
          - 'null'
          format: uuid
        upper_storage_object_id:
          type:
          - string
          - 'null'
          format: uuid
          description: |-
            For OverlayBD disks: ID of the OverlaybdUpper StorageObject holding the
            persistent upper layer (upper.data + upper.index) on a Local or NFS pool.
            None = ephemeral (deleted on VM stop); Some = persistent across VM delete.
        vhost_socket:
          type:
          - string
          - 'null'
        vhost_user:
          type: boolean
        vm_id:
          type: string
          format: uuid
...
    VmImagePreflightRequest:
      type: object
      required:
      - image_ref
      properties:
        architecture:
          type:
          - string
          - 'null'
        boot_mode:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/BootMode'
        host_id:
          type:
          - string
          - 'null'
          format: uuid
        image_ref:
          type: string
    VmImagePreflightResponse:
      type: object
      required:
      - bootable
      - host_id
      - host_name
      - resolved_image_ref
      - architecture
      - checks
      properties:
        architecture:
          type: string
        bootable:
          type: boolean
        checks:
          type: array
          items:
            $ref: '#/components/schemas/VmImagePreflightCheck'
        host_id:
          type: string
          format: uuid
        host_name:
          type: string
        resolved_image_ref:
          type: string
...
    VmMigrateRequest:
      type: object
      description: Request body for `POST /vms/{vm_id}/migrate`.
      required:
      - target_host_id
      properties:
        target_host_id:
          type: string
          format: uuid
          description: UUID of the destination host to migrate the VM to.
...